refactor(utils): extract createParserErrorCapturer and ParserErrorCapturer in a shared file
