+++ packages/react-native-codegen/src/parsers/typescript/index.js
@@ -13,17 +13,16 @@
 import type {SchemaType} from '../../CodegenSchema.js';
 const babelParser = require('@babel/parser');
 const fs = require('fs');
-const {extractNativeModuleName} = require('../utils.js');
+const {
+  extractNativeModuleName,
+  createParserErrorCapturer,
+} = require('../utils.js');
 const {buildComponentSchema} = require('./components');
 const {wrapComponentSchema} = require('./components/schema');
 const {buildModuleSchema} = require('./modules');
 const {wrapModuleSchema} = require('../parsers-commons');
 
-const {
-  createParserErrorCapturer,
-  visit,
-  isModuleRegistryCall,
-} = require('./utils');
+const {visit, isModuleRegistryCall} = require('./utils');
 const invariant = require('invariant');
 
 function getConfigType(
